examples/simple/requirements.txt
djangoapi-guard
gunicorn
redis
